Remove tighter restrictions on XLink arc element
The current definition made from and to attributes mandatory, going
beyond the XLink specification, which lets missing from and to
attributes devolve to meaning all locators. Since this is also fine for
the purposes of SSP Traceability and concurs with the optional label for
locators, this additional restriction can be removed. Fixes #55.